// frontend/src/app/generated/model/users.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Users {
    createdOn?: Date;
    lastUpdateOn?: Date;
    name?: string;
    password?: string;
    permissionList?: Array<string>;
    userRole?: string;
    status?: number;
    surname?: string;
    userEmail?: string;
    userId?: number;
    username?: string;
}